chore: add logging, refactor module, fix missing parameters in default values
