+++ .env.template
@@ -1,5 +1,3 @@
 ROOT_PATH=/eidos/
-API_HOST=127.0.0.1
-API_PORT=6004
-LOG_LEVEL=info
+API_KEY=API_KEY_GOES_HERE
 FUNCTIONS_FOLDER=functions
